Добавляет новые модули
